% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellatac_utils.R
\name{annotate_gr}
\alias{annotate_gr}
\title{Annotate GenomicRanges}
\usage{
annotate_gr(gr, EnsDb.genome = EnsDb.Hsapiens.v86,
  blacklist_gr = Signac::blacklist_hg38, minoverlap = 10)
}
\arguments{
\item{gr}{GenomicRanges object to annotate}

\item{EnsDb.genome}{Ensembl annotation package to use for annotation. See details for installation (default = EnsDb.Hsapiens.v86, for hg38).}

\item{blacklist_gr}{GenomicRanges object of ENCODE blacklist regions (default: blacklist_hg38 from Signac package)}

\item{minoverlap}{minimum no of overlapping bps to call an overlap}
}
\description{
add width and overlap with annotations to metadata columns of GenomicRanges object
}
\details{
You can install EnsDb annotation packages by running
\code{BiocManager::install("EnsDb.Hsapiens.v86")}
}
