% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellatac_utils.R
\name{compute_peakQC}
\alias{compute_peakQC}
\title{Compute QC metric for peaks}
\usage{
compute_peakQC(win_seu, cluster_col = "seurat_clusters",
  EnsDb.genome = EnsDb.Hsapiens.v86, blacklist_gr = blacklist_hg38)
}
\arguments{
\item{win_seu}{Seurat object}

\item{cluster_col}{column in metadata storing cluster information}

\item{EnsDb.genome}{Ensembl annotation package of genome of interest (default = EnsDb.Hsapiens.v86)}

\item{blacklist_gr}{GenomicRanges object of ENCODE blacklisted regions for genome of interest
(default uses blacklist annotation implemented in Signac)}
}
\value{
Seurat object with qc metric stored in `meta.features` of `peaks` assay
}
\description{
Compute QC metric for peaks
}
