% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellatac_utils.R
\name{fracAccessible_cluster}
\alias{fracAccessible_cluster}
\title{Calculate accessible fraction per cluster}
\usage{
fracAccessible_cluster(win_seu, assay, group_col = "seurat_clusters")
}
\arguments{
\item{win_seu}{Seurat object}

\item{assay}{assay in seurat object to use}

\item{group_col}{meta.data column storing cluster identity (or other grouping variable for cells)}
}
\value{
matrix of features x clusters storing fractions
}
\description{
For each feature it computes the fraction of cells per cluster in which that feature is accessible
}
