% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellatac_utils.R
\name{get_annotation_overlap}
\alias{get_annotation_overlap}
\title{Get annotation overlap}
\usage{
get_annotation_overlap(gr, anno_gr, anno_name, minoverlap = 10)
}
\arguments{
\item{gr}{GenomicRanges object}

\item{anno_gr}{GenomicRanges object of annotations (usually loaded from ensembldb package)}

\item{anno_name}{character of name of annotation (e.g. genes, exons, promoters ...)}

\item{minoverlap}{minimum no of overlapping bps to call an overlap}
}
\value{
GenomicRanges object with overlap stored in metadata column
}
\description{
Computes overlap between GenomicRanges object and annotation coordinates
}
