% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellatac_utils.R
\name{load_cellatac_peakmat}
\alias{load_cellatac_peakmat}
\title{Load peaks x cell matrix}
\usage{
load_cellatac_peakmat(out.dir, as.Seurat = TRUE)
}
\arguments{
\item{out.dir}{path of cellatac output directory}

\item{as.Seurat}{logical indicating whether output should be a SeuratAssay object. If FALSE, a sparseMatrix is returned (default: TRUE)}
}
\value{
peak matrix (SeuratAssay object or sparseMatrix)
}
\description{
Load peaks x cell matrix
}
