% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellatac_utils.R
\name{load_cellatac_seurat}
\alias{load_cellatac_seurat}
\title{Load cellatac output as Seurat object}
\usage{
load_cellatac_seurat(out.dir)
}
\arguments{
\item{out.dir}{path of cellatac output directory}
}
\value{
SeuratObject with 2 assays (windows and peaks)
}
\description{
Load cellatac output as Seurat object
}
